refactor removing outdated events in util function
