+++ src/controller/main/EventsController.ts
@@ -3,7 +3,7 @@
 
 import { getUid } from '@/utils/CryptoUtils';
 import { MainDebug } from '@/utils/DebugUtils';
-import { pushUniqueEvent } from '@/utils/EventUtils';
+import { doRemoveOutdatedEvents, pushUniqueEvent } from '@/utils/EventUtils';
 import { store } from '@/main';
 import { WindowsController } from '@/controller/main/WindowsController';
 import type { AnyJson } from '@w3ux/utils/types';
@@ -143,44 +143,9 @@ export class EventsController {
    * Will remove old matching events from the store.
    */
   static removeOutdatedEvents(event: EventCallback) {
-    switch (event.taskAction) {
-      case 'subscribe:account:nominationPools:rewards':
-      case 'subscribe:account:nominating:pendingPayouts': {
-        const { address, chainId } = event.who.data as EventAccountData;
-        const { taskAction } = event;
-        const all = EventsController.getEventsFromStore();
-
-        const updated = all.filter((ev) => {
-          // Keep if it's a chain event.
-          if (ev.who.origin === 'chain') {
-            return true;
-          }
-
-          // Extract target data from next event.
-          const { taskAction: nextTaskAction } = ev;
-          const { address: nextAddress, chainId: nextChainId } = ev.who
-            .data as EventAccountData;
-
-          // Remove event if its task action, address and chain id are the same.
-          if (
-            nextTaskAction === taskAction &&
-            nextAddress === address &&
-            nextChainId === chainId
-          ) {
-            return false;
-          }
-
-          // Otherwise, keep the event.
-          return true;
-        });
-
-        this.persistEventsToStore(updated);
-        break;
-      }
-      default: {
-        break;
-      }
-    }
+    const all = EventsController.getEventsFromStore();
+    const { updated, events } = doRemoveOutdatedEvents(event, all);
+    updated && this.persistEventsToStore(events);
   }
 
   /**
