+++ src/renderer/contexts/main/Events/index.tsx
@@ -1,15 +1,15 @@
 // Copyright 2024 @rossbulat/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import React, { createContext, useContext, useState } from 'react';
-import { pushUniqueEvent, getEventChainId } from '@/utils/EventUtils';
 import * as defaults from './defaults';
+import React, { createContext, useContext, useState } from 'react';
+import {
+  pushUniqueEvent,
+  getEventChainId,
+  doRemoveOutdatedEvents,
+} from '@/utils/EventUtils';
 import type { ChainID } from '@/types/chains';
-import type {
-  DismissEvent,
-  EventAccountData,
-  EventCallback,
-} from '@/types/reporter';
+import type { DismissEvent, EventCallback } from '@/types/reporter';
 import type {
   EventsContextInterface,
   EventsState,
@@ -43,46 +43,22 @@ export const EventsProvider = ({ children }: { children: React.ReactNode }) => {
 
   /// Remove any outdated events in the state.
   const removeOutdatedEvents = (event: EventCallback) => {
-    const { taskAction } = event;
-    const { address } = event.who.data as EventAccountData;
+    setEventsState((prev) => {
+      const cloned = new Map(prev);
+      const chainId = getEventChainId(event);
 
-    switch (taskAction) {
-      case 'subscribe:account:nominationPools:rewards':
-      case 'subscribe:account:nominating:pendingPayouts': {
-        setEventsState((prev) => {
-          const cloned = new Map(prev);
-          const chainId = getEventChainId(event);
-          let curEvents = cloned.get(chainId);
-
-          // Filter outdated events if any exist.
-          if (curEvents !== undefined) {
-            curEvents = curEvents.filter((ev) => {
-              if (ev.who.origin === 'chain') {
-                return true;
-              }
-
-              // Extract target data from next event.
-              const { taskAction: nextTaskAction } = ev;
-              const { address: nextAddress } = ev.who.data as EventAccountData;
-
-              // Remove event if its task action and address are the same.
-              if (nextTaskAction === taskAction && nextAddress === address) {
-                return false;
-              }
-
-              // Otherwise, keep the event.
-              return true;
-            });
-            cloned.set(chainId, curEvents);
-          }
-          return cloned;
-        });
-        break;
+      const all = cloned.get(chainId);
+      if (!all) {
+        return cloned;
       }
-      default: {
-        break;
+
+      const { updated, events: newEvents } = doRemoveOutdatedEvents(event, all);
+      if (updated) {
+        cloned.set(chainId, newEvents);
       }
-    }
+
+      return cloned;
+    });
   };
 
   /// Adds an event to the events state.
